+++ src/draw/state.rs
@@ -1,10 +1,13 @@
+#![allow(clippy::cast_possible_truncation)]
 use crate::{command::DataCmd, comms::Response, ScreenPosition, Turtle};
 
 /*
  * Tell Turtle's State
  */
 impl Turtle {
-    pub fn position(&self) -> ScreenPosition<isize> {
+    /// # Panics
+    /// Panics when there's a library bug
+    pub fn position(&self) -> ScreenPosition<i32> {
         if let Response::Position(pos) = self.do_data(DataCmd::Position) {
             [pos.x, -pos.y].into()
         } else {
@@ -12,10 +15,12 @@ impl Turtle {
         }
     }
 
-    pub fn pos(&self) -> ScreenPosition<isize> {
+    pub fn pos(&self) -> ScreenPosition<i32> {
         self.position()
     }
 
+    /// # Panics
+    /// Panics when there's a library bug
     pub fn towards<X: Into<f64>, Y: Into<f64>>(&self, xpos: X, ypos: Y) -> f32 {
         let x = xpos.into() as f32;
         let y = ypos.into() as f32;
@@ -26,14 +31,16 @@ impl Turtle {
         }
     }
 
-    pub fn xcor(&self) -> isize {
+    pub fn xcor(&self) -> i32 {
         self.position().x
     }
 
-    pub fn ycor(&self) -> isize {
+    pub fn ycor(&self) -> i32 {
         self.position().y
     }
 
+    /// # Panics
+    /// Panics when there's a library bug
     pub fn heading(&self) -> f32 {
         if let Response::Heading(angle) = self.do_data(DataCmd::Heading) {
             angle
@@ -42,12 +49,12 @@ impl Turtle {
         }
     }
 
-    pub fn distance<D: Into<ScreenPosition<isize>>>(&self, other: D) -> f64 {
+    pub fn distance<D: Into<ScreenPosition<i32>>>(&self, other: D) -> f64 {
         let self_pos = self.pos();
-        let other_pos: ScreenPosition<isize> = other.into();
+        let other_pos: ScreenPosition<i32> = other.into();
 
-        let dx = (other_pos.x - self_pos.x) as f64;
-        let dy = (other_pos.y - self_pos.y) as f64;
+        let dx = f64::from(other_pos.x - self_pos.x);
+        let dy = f64::from(other_pos.y - self_pos.y);
 
         (dx * dx + dy * dy).sqrt()
     }
